pyproject.toml
[project]
name = "data-pipeline-prod"
version = "0.1.0"
description = ""
authors = [
    {name = "Mairon Costa",email = "duartemairon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "faker (>=40.1.0,<41.0.0)"
]
